Fix incompatibility with OpenSSL 1.1
The pre-compiled PhantomJS binary used in this package dates from 2014 and is
linked to use libcrypto.so and libssl.so from OpenSSL 1.0, but does not
specify their version (e.g. libcrypto.so.1.0.0). Hence, running PhantomJS
crashes (segmentation fault) on systems with OpenSSL 1.1 installed, because
libcrypto.so and libssl.so refer to the new version 1.1. Removing
/lib64/libssl.so from the system removes the crash, but of course it's not a
real fix. I've tried multiple solutions including using LD_PRELOAD or
patchelf, but the only working I've found is to alter PhantomJS binary to make
it fail loading libcrypto.so and libssl.so.

Tested on Fedora 32 and CentOS 7.